fonction en int

+++ main.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 #include <stdlib.h>
 
-double puissance(double x, double n);
+int puissance(int x, int n);
 
 int main()
 {
@@ -16,7 +16,7 @@ int main()
     return 0;
 }
 
-double puissance(double x, double n){
+int puissance(int x, int n){
     if(n == 0) return 1;
     return (puissance(x,(n-1)) * x);
 }
